TEMP: prepare package structure for merging.
